refactor(web): split event-parser.ts (863→351 lines) into focused sub-modules
Extract 5 sub-parser modules from the monolithic parseLegacyEvent switch:

- parse-interaction-events.ts (150 lines) — secret, confirmation,
  contact, question request parsers
- parse-tool-events.ts (112 lines) — tool_use_start, tool_result,
  tool_progress parsers
- parse-surface-events.ts (89 lines) — ui_surface_show/update/
  dismiss/complete parsers
- parse-subagent-events.ts (108 lines) — subagent spawn, status,
  inner event parsers
- parse-resource-events.ts (198 lines) — sync_changed, identity,
  avatar, conversation title, notification, disk pressure,
  document editor parsers

event-parser.ts remains the thin orchestrator: envelope unwrap →
canonical schema dispatch → legacy switch delegation.

Also:
- Extract parse-helpers.ts with shared unknownEvent() (was
  duplicated 3x)
- Move toDisplayAttachments to utils/display-attachments.ts (pure
  function, not event parsing — was misplaced in lib/)
- Move toDisplayAttachments tests to colocated test file
- Update stream-message-updaters.ts import path

Zero behavioral changes — all 127 tests pass unchanged.

Closes LUM-2055
