+++ apps/web/src/utils/display-attachments.test.ts
@@ -1,0 +1,84 @@
+import { describe, expect, test } from "bun:test";
+
+import { toDisplayAttachments } from "@/utils/display-attachments";
+
+describe("toDisplayAttachments", () => {
+  test("returns undefined for empty/missing input", () => {
+    expect(toDisplayAttachments(undefined)).toBeUndefined();
+    expect(toDisplayAttachments([])).toBeUndefined();
+  });
+
+  test("converts image attachment with data-URI previewUrl", () => {
+    const result = toDisplayAttachments([
+      {
+        id: "att-1",
+        filename: "photo.png",
+        mimeType: "image/png",
+        data: "iVBORw0KGgo=",
+      },
+    ]);
+    expect(result).toEqual([
+      {
+        id: "att-1",
+        filename: "photo.png",
+        mimeType: "image/png",
+        sizeBytes: expect.any(Number),
+        previewUrl: "data:image/png;base64,iVBORw0KGgo=",
+      },
+    ]);
+  });
+
+  test("creates data-URI previewUrl for non-image types with inline data", () => {
+    const result = toDisplayAttachments([
+      {
+        id: "att-2",
+        filename: "report.pdf",
+        mimeType: "application/pdf",
+        data: "JVBERi0xLjQ=",
+      },
+    ]);
+    expect(result).toEqual([
+      {
+        id: "att-2",
+        filename: "report.pdf",
+        mimeType: "application/pdf",
+        sizeBytes: expect.any(Number),
+        previewUrl: "data:application/pdf;base64,JVBERi0xLjQ=",
+      },
+    ]);
+  });
+
+  test("uses thumbnailData when data is empty", () => {
+    const result = toDisplayAttachments([
+      {
+        id: "att-3",
+        filename: "clip.mp4",
+        mimeType: "video/mp4",
+        data: "",
+        thumbnailData: "thumb123",
+        fileBacked: true,
+        sizeBytes: 1024,
+      },
+    ]);
+    expect(result).toEqual([
+      {
+        id: "att-3",
+        filename: "clip.mp4",
+        mimeType: "video/mp4",
+        sizeBytes: 1024,
+        previewUrl: "data:image/jpeg;base64,thumb123",
+      },
+    ]);
+  });
+
+  test("falls back to filename for id when id is missing", () => {
+    const result = toDisplayAttachments([
+      {
+        filename: "noId.txt",
+        mimeType: "text/plain",
+        data: "aGVsbG8=",
+      },
+    ]);
+    expect(result?.[0]?.id).toBe("noId.txt");
+  });
+});
